pyproject.toml
[tool.poetry]
name = "telegram-bot-template"
version = "0.1.0"
description = ""
authors = ["Fobos <95380054+Yastrah@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.18.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.5"
sqlalchemy = "^2.0.38"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
